requirements.txt
numpy~=1.19.5
pytorch3d~=0.2.5
scikit-image~=0.17.2
wandb~=0.10.30
opencv-python~=4.5.1.48
scipy~=1.5.4
resampy~=0.2.2
torchvision~=0.8.1
torchfile~=0.1.0
omegaconf~=2.0.6
adabound~=0.0.5
smplx~=0.1.26
imageio~=2.9.0
pyvista~=0.26.1
tqdm~=4.51.0
pillow~=8.1.2
pandas~=1.1.4
imgaug~=0.4.0
scikit-learn~=0.23.2
setuptools~=54.2.0
matplotlib~=3.3.4
psutil~=5.7.3
pyyaml~=5.4.1
configparser~=5.0.1